feat: 表格溢出 tooltip 能力公共化下沉为 trace 公共 hook 并接入主机/进程列表 --story=136776188 # Reviewed, transaction id: 83812
