+++ bkmonitor/webpack/src/trace/pages/trace-explore/components/trace-explore-table/utils/index.ts
@@ -25,4 +25,3 @@
  */
 
 export * from './api-utils';
-export * from './dom-helper';
