+++ bkmonitor/webpack/src/trace/utils/index.ts
@@ -23,5 +23,6 @@
  * CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
  * IN THE SOFTWARE.
  */
+export * from './dom-helper';
 export * from './utils';
 export * from './variable';
